requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.5
pymongo==4.5.0
python-dotenv==1.0.0
cloudinary==1.34.0
gunicorn==21.2.0
eventlet==0.33.3
dnspython==2.4.2
gevent==23.9.1
gevent-websocket==0.10.1